Moved components to seperate files
