+++ client/components/index.js
@@ -1,0 +1,7 @@
+import Button from './Button'
+import Card from './Card'
+import Input from './Input'
+import List from './List'
+// import Paper from './Paper'
+
+export { Button, Card, Input, List }
